Aient

+++ controllers/admin.js
@@ -1,8 +1,9 @@
+const path = require('path');
+const fileHelper = require('../util/file');
 const Product = require('../models/product');
 const {
   validationResult
 } = require('express-validator');
-const path = require('path');
 
 exports.getAddProduct = (req, res, next) => {
   res.render('admin/edit-product', {
@@ -57,10 +58,7 @@ exports.postAddProduct = (req, res, next) => {
       validationErrors: errors.array()
     });
   }
-  console.log(image);
   const imageUrl = path.normalize(('/' + image.path));
-  console.log(imageUrl);
-
   const product = new Product({
     title: title,
     price: price,
@@ -167,7 +165,9 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if (image) {
-        product.imageUrl = image.path.replace("\\", "/");
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = path.normalize(('/' + image.path));
+
       }
       return product.save().then(result => {
         console.log('UPDATED PRODUCT!');
@@ -204,11 +204,19 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({
-      _id: prodId,
-      userId: req.user._id
-    })
-    .then(() => {
+
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error('Product not found'));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+
+      return Product.deleteOne({
+        _id: prodId,
+        userId: req.user._id
+      });
+    }).then((result) => {
       res.redirect('/admin/products');
     })
     .catch(err => {
